moved Cockpit stats hydration files from utils to liveapp folder, modified files to reflect those changes
