+++ lib/liveApp/cockpitStatsHydration.js
