+++ lib/socket/handlers.js
@@ -29,7 +29,7 @@ const coroutine = Promise.coroutine;
 const settings = require('../settings').presentationSettings;
 const presUtils = require('../utils').presentation;
 const pluginPubsub = require('../plugin/pubsub');
-const liveApp = require('../liveApp');
+const liveApp = require('../liveApp').liveApp;
 
 module.exports =  function(socketUtils){
