+++ lib/socket/liveApp.js
@@ -1,6 +1,6 @@
  // lib/socket/liveApp.js
 
-const liveApp = require('../liveApp')
+const liveApp = require('../liveApp').liveApp;
 
  
  module.exports = {
