+++ lib/socket/presenterHandlers.js
@@ -30,7 +30,7 @@ const settings = require('../settings').presentationSettings;
 const presUtils = require('../utils').presentation;
 const pluginPubsub = require('../plugin/pubsub');
 const renameIdAndRemoveVersion =  require('../utils/format').renameIdAndRemoveVersion;
-const cockpitStatsHydration = require('../utils').cockpitStatsHydration
+const cockpitStatsHydration = require('../liveApp').cockpitStatsHydration
 
 
 module.exports =  function(socketUtils){
