+++ lib/utils/index.js
@@ -7,5 +7,4 @@ module.exports = {
   presentation   : require('./presentation'),
   routes         : require('./routes'),
   stats          : require('./stats'),
-  cockpitStatsHydration : require('./cockpitStatsHydration')
 }
